--- 07week/API.html
@@ -1,23 +1,0 @@
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
-		<title>Fetch Address Book</title>
-	</head>
-	<body>
-		<div id="container">
-			<button onclick="getAPI()">Get API</button>
-			<div></div>
-			<h3>All Posts</h3>
-			<ul id="all-posts"></ul>
-		</div>
-		<script src="./api.js"></script>
-	</body>
-</html>
-
-
-
-http://api.brewerydb.com/v2/{endpoint}/?key=abcdef.
-
-https://sandbox-api.brewerydb.com/v2/{endpoint}/?key=2c035d98e35443476d5d06e2f9f62d19.
