+++ 08week/Dodgeball/Dodgeball.html
@@ -18,12 +18,14 @@
 		</div>
 		<div>
 			<h4>Blue Team</h4>
+
 			<ul id="blue"></ul>
 		</div>
 		<div>
 			<h4>Red Team</h4>
 			<ul id="red"></ul>
 		</div>
-		<script src="./dodgeball.js"></script>
+
+		<script src="./main.js"></script>
 	</body>
 </html>
