--- 08week/fetch/index.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<title>Fetch</title>
-		<link
-			rel="stylesheet"
-			href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css"
-		/>
-		<link rel="stylesheet" href="style.css" />
-	</head>
-	<body>
-		<div id="fetch"></div>
-		<script
-			type="text/javascript"
-			src="https://cdnjs.cloudflare.com/ajax/libs/babel-standalone/6.23.1/babel.min.js"
-		></script>
-		<script
-			type="text/javascript"
-			src="https://cdnjs.cloudflare.com/ajax/libs/react/15.5.0/react.js"
-		></script>
-		<script
-			type="text/javascript"
-			src="https://cdnjs.cloudflare.com/ajax/libs/react/15.5.0/react-dom.js"
-		></script>
-		<script type="text/babel" src="./script.js"></script>
-		<script>
-			// let first = new Promise((resolve, reject) => {
-			// 	resolve("Max");
-			// });
-			// let second = dataFromFirst => {
-			// 	return new Promise((resolve, reject) => {
-			// 		resolve(dataFromFirst + " is awesome");
-			// 	});
-			// };
-			// //console.log(second());
-			// first
-			// 	.then(data => {
-			// 		return second(data);
-			// 	})
-			// 	.then(data => {
-			// 		console.log(data);
-			// 	});
-		</script>
-		<button id="getAPI">Get API</button>
-	</body>
-	<!-- <button id="get">Get text</button>
-	<button id="getJson">Get text</button> -->
-</html>
