--- 09week/hackernews/pokemon.html
@@ -1,13 +1,0 @@
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
-		<title>Fetch Address Book</title>
-	</head>
-	<body>
-		<h3>Pokemon</h3>
-
-		<script src="./pokemon.js"></script>
-	</body>
-</html>
